% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bin_data.R
\name{convert_spe_to_image}
\alias{convert_spe_to_image}
\title{Convert a Spatial Experiment object to a rasterized image}
\usage{
convert_spe_to_image(spe, nbins_x = NBINS, nbins_y = NBINS, nas_to_zero = TRUE)
}
\arguments{
\item{spe}{A \link[SpatialExperiment:SpatialExperiment]{SpatialExperiment::SpatialExperiment} object containing the
data to rasterize.}

\item{nbins_x}{Number of bins to use on the X axis (default is 300).}

\item{nbins_y}{Number of bins to use on the Y axis (default is 300).}

\item{nas_to_zero}{Whether or not to convert missing values to 0 (default is
\code{TRUE}).}
}
\value{
An array of dimensions \code{nbins_x} by \code{nbins_y} containing the binned
counts for each "pixel".
}
\description{
This function bins the total counts and returns a rasterized image.
}
