% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{load_as_spatial_experiment}
\alias{load_as_spatial_experiment}
\title{Load a counts file and metadata file as a SpatialExperiment object}
\usage{
load_as_spatial_experiment(
  counts_file,
  metadata_file,
  metadata = NULL,
  id_columns = c("fov", "cell_ID"),
  coord_names = c("CenterX_global_px", "CenterY_global_px"),
  ...
)
}
\arguments{
\item{counts_file}{Path to the file containing the feature counts per
spot/cell.}

\item{metadata_file}{Path to the file containing the spot/cell metadata. Must
include the spatial coordinates.}

\item{metadata}{Object containing the spot/cell metadata. If one is supplied,
\code{metadata_file} is ignored.}

\item{id_columns}{Names of the columns in common between \code{counts} and
\code{metadata}, used for aligning the spots/cells. Default is
\code{c("fov", "cell_ID")}.}

\item{coord_names}{A vector of length 2, containing the names of the columns
containing the x and y coordinates in the metadata. Default is
\code{c("CenterX_global_px", "CenterY_global_px")} (which are the default
Cosmx column names).}

\item{...}{Additional parameters passed to the
\link[SpatialExperiment:SpatialExperiment]{SpatialExperiment::SpatialExperiment} constructor.}
}
\value{
A \code{SpatialExperiment} object.
}
\description{
\code{load_as_spatial_experiment} loads the corresponding files from disk,
combines them based on the provided identifying columns and returns a
SpatialExperiment object.
}
