% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_counts}
\alias{plot_counts}
\title{Plots the image produced by rasterizing the counts data.}
\usage{
plot_counts(
  counts_image,
  flip = FALSE,
  points = data.frame(),
  log1p = FALSE,
  minimal = FALSE
)
}
\arguments{
\item{counts_image}{The array of binned counts, as returned by
\link{convert_spe_to_image}.}

\item{flip}{Whether or not to reverse the x and/or y axes. Should be one of
"x", "X", "y", "Y", \code{TRUE}, or \code{FALSE} (default).}

\item{points}{Optional data.frame containing the coordinates of points to
plot on the image.}

\item{log1p}{If \code{TRUE}, the counts are normalized prior to plotting. Default
is \code{FALSE}.}

\item{minimal}{Whether to remove the axis text and legend. Default is
\code{FALSE}.}
}
\value{
A ggplot object.
}
\description{
Plots the image produced by rasterizing the counts data.
}
