% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_register.R
\name{shiny_register}
\alias{shiny_register}
\title{Launch the shiny interface to align two spatial transcriptomics samples}
\usage{
shiny_register(source, reference, nbins = 200)
}
\arguments{
\item{source}{A \link[SpatialExperiment:SpatialExperiment]{SpatialExperiment::SpatialExperiment} object containing the
sample to rotate.}

\item{reference}{A \link[SpatialExperiment:SpatialExperiment]{SpatialExperiment::SpatialExperiment} object containing
the reference sample.}

\item{nbins}{The number of bins on each axis. This is mainly for the image
display for the definition of the points and has no impact on the final
object.}
}
\value{
A \link[SpatialExperiment:SpatialExperiment]{SpatialExperiment::SpatialExperiment} object containing the
rotated sample with the new coordinates.
}
\description{
This launches a shiny interface displaying a rasterized version of the
samples (total counts per bin). Three points must be placed on the image of
the sample to rotate, with their corresponding projections on the reference
image. The coordinates of the first sample are then transformed such that the
points are lined up in both. The affine matrix and the resulting
superposition of the samples are displayed. When the result is satisfactory,
the \code{SpatialExperiment} object with the transformed coordinates can be saved
as an \code{RDS} file.
}
